pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ds-churn"
version = "1.0.0"
description = "Churn Prediction — Modular Monolith ML pipeline"
requires-python = ">=3.10"
readme = "README.md"

dependencies = [
    # --- Data Processing ---
    "pandas>=2.2,<3",
    "numpy>=1.26,<2",

    # --- Database ---
    "psycopg2-binary>=2.9,<3",
    "sqlalchemy>=2.0,<3",

    # --- ML ---
    "scikit-learn>=1.5,<2",
    "xgboost>=2.0",
    "joblib>=1.3",

    # --- Utilities ---
    "python-dotenv>=1.0,<2",
    "python-dateutil>=2.8",
    "pydantic>=2.5,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.4",
]
scheduler = [
    "apscheduler>=3.10,<4",
]

# ---------------------------------------------------------------------------
# Package discovery — src layout
# ---------------------------------------------------------------------------
[tool.setuptools.packages.find]
where = ["src"]

# ---------------------------------------------------------------------------
# pytest
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
addopts = "-v --tb=short"

# ---------------------------------------------------------------------------
# ruff (linter + formatter)
# ---------------------------------------------------------------------------
[tool.ruff]
src = ["src"]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "S",   # flake8-bandit (security)
    "T20", # flake8-print
]
ignore = [
    "S101",  # assert used (OK in tests)
    "T201",  # print (OK in CLI scripts)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101", "T201"]

[tool.ruff.lint.isort]
known-first-party = ["config", "shared", "data", "modeling", "monitoring", "features", "pipelines"]
